search_channel_videos/requirements.txt
google-cloud-storage==1.35.0
google-api-python-client==1.12.8
google-auth==1.24.0
PyYAML==5.3.1